fixed compile errors and warningsa
<==== Mercurial data ====>
Branch: dev_3.0.0

+++ common_libs/nx_network/src/nx/network/test_support/simple_socket_test_helper.h
@@ -253,7 +253,7 @@ void socketSimpleAsync(
             {
                 if (code == SystemError::noError)
                 {
-                    EXPECT_GT(size, 0);
+                    EXPECT_GT(size, (size_t)0);
                     EXPECT_STREQ(serverBuffer.data(), testMessage.data());
                     serverBuffer.resize(0);
                 }
@@ -286,7 +286,7 @@ void socketSimpleAsync(
                         return;
 
                     EXPECT_EQ(code, SystemError::noError);
-                    EXPECT_EQ(size, testMessage.size());
+                    EXPECT_EQ(size, (size_t)testMessage.size());
                 });
         });
